+++ previous-versions/assess/pkg/models/operations/getloansummary.go
@@ -8,6 +8,7 @@ import (
 )
 
 type GetLoanSummaryRequest struct {
+	// Unique identifier for a company.
 	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
 }
 
@@ -19,12 +20,15 @@ func (o *GetLoanSummaryRequest) GetCompanyID() string {
 }
 
 type GetLoanSummaryResponse struct {
+	// HTTP response content type for this operation
 	ContentType string
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
 	// OK
 	LoanSummary *shared.LoanSummary
-	StatusCode  int
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
 	RawResponse *http.Response
 }